+++ quake3-rs/cgamex86_64/src/cgame/cg_event.rs
@@ -488,7 +488,7 @@ pub unsafe extern "C" fn CG_PlaceString(mut rank: i32) -> *const libc::c_char {
     }
     Com_sprintf(
         str.as_mut_ptr(),
-        ::std::mem::size_of::<[libc::c_char; 64]>() as libc::c_ulong as i32,
+        ::std::mem::size_of::<[libc::c_char; 64]>() as usize as i32,
         b"%s%s\x00" as *const u8 as *const libc::c_char,
         t,
         s,
@@ -533,8 +533,8 @@ unsafe extern "C" fn CG_Obituary(mut ent: *mut entityState_t) {
     Q_strncpyz(
         targetName.as_mut_ptr(),
         Info_ValueForKey(targetInfo, b"n\x00" as *const u8 as *const libc::c_char),
-        (::std::mem::size_of::<[libc::c_char; 32]>() as libc::c_ulong)
-            .wrapping_sub(2 as i32 as libc::c_ulong) as i32,
+        (::std::mem::size_of::<[libc::c_char; 32]>() as usize)
+            .wrapping_sub(2 as i32 as usize) as i32,
     );
     libc::strcat(
         targetName.as_mut_ptr(),
@@ -662,8 +662,8 @@ unsafe extern "C" fn CG_Obituary(mut ent: *mut entityState_t) {
         Q_strncpyz(
             attackerName.as_mut_ptr(),
             Info_ValueForKey(attackerInfo, b"n\x00" as *const u8 as *const libc::c_char),
-            (::std::mem::size_of::<[libc::c_char; 32]>() as libc::c_ulong)
-                .wrapping_sub(2 as i32 as libc::c_ulong) as i32,
+            (::std::mem::size_of::<[libc::c_char; 32]>() as usize)
+                .wrapping_sub(2 as i32 as usize) as i32,
         );
         libc::strcat(
             attackerName.as_mut_ptr(),
@@ -674,7 +674,7 @@ unsafe extern "C" fn CG_Obituary(mut ent: *mut entityState_t) {
             Q_strncpyz(
                 cg.killerName.as_mut_ptr(),
                 attackerName.as_mut_ptr(),
-                ::std::mem::size_of::<[libc::c_char; 32]>() as libc::c_ulong as i32,
+                ::std::mem::size_of::<[libc::c_char; 32]>() as usize as i32,
             );
         }
     }
